Set up automatic package publishing and rename SDK
In order to streamline new version releases, a GitHub workflow for automatically publishing the package to PyPi when a new release is made has been added. Additionally, the name of the SDK has been changed in setup.py from "SkipCash Python SDK" to "skipcash" to match the package name in PyPi.+.gitignore was also updated to ignore new egg-info folder and distribution files.