Update PyPI publish action version
The GitHub Actions workflow for publishing packages to PyPI has been updated to use a more recent version of the action. The PyPI publish action version has been updated from release/v1 to release/v1.8 to ensure the most up-to-date features and bug fixes are utilized in the publishing process.